+++ apps/backend-functions/src/internals/users.ts
@@ -6,8 +6,18 @@ import { userInvite, userFirstConnexion } from '../templates/mail';
 import { groupIds, templateIds } from '@blockframes/utils/emails/ids';
 import { auth, db } from './firebase';
 import { sendMailFromTemplate, sendMail } from './email';
-import { OrganizationDocument, PublicOrganization, InvitationMode, InvitationStatus, InvitationType, PublicUser, App } from '@blockframes/model';
-import { EventEmailData, getOrgEmailData, getUserEmailData } from '@blockframes/utils/emails/utils';
+import {
+  EventEmailData,
+  getOrgEmailData,
+  getUserEmailData,
+  OrganizationDocument,
+  PublicOrganization,
+  InvitationMode,
+  InvitationStatus,
+  InvitationType,
+  PublicUser,
+  App,
+} from '@blockframes/model';
 import { logger } from 'firebase-functions';
 import { hasUserAnOrgOrIsAlreadyInvited } from '../invitation';
