+++ apps/backend-functions/src/templates/mail.ts
@@ -4,7 +4,6 @@
  * https://www.notion.so/cascade8/Email-Data-Object-8ed9d64e8cd4490ea7bc0e469c04043e
  */
 import { supportEmails, appUrl, e2eMode } from '../environments/environment';
-import { EmailRequest, EmailTemplateRequest } from '../internals/email';
 import { templateIds } from '@blockframes/utils/emails/ids';
 import { RequestDemoInformations } from '@blockframes/utils/request-demo';
 import {
@@ -19,9 +18,9 @@ import {
   Timestamp,
   App,
   appName,
-  Module
-} from '@blockframes/model';
-import {
+  Module,
+  EmailRequest,
+  EmailTemplateRequest,
   EventEmailData,
   OrgEmailData,
   UserEmailData,
@@ -29,8 +28,8 @@ import {
   getOfferEmailData,
   MovieEmailData,
   getBucketEmailData,
-  getNegotiationEmailData
-} from '@blockframes/utils/emails/utils';
+  getNegotiationEmailData,
+} from '@blockframes/model';
 import { format } from "date-fns";
 import { supportMailosaur } from '@blockframes/utils/constants';
