+++ apps/backend-functions/src/users.ts
@@ -11,9 +11,18 @@ import { getMailSender, applicationUrl } from '@blockframes/utils/apps';
 import { sendFirstConnexionEmail, createUserFromEmail } from './internals/users';
 import { production } from './environments/environment';
 import { cleanUserMedias } from './media';
-import { getUserEmailData, OrgEmailData } from '@blockframes/utils/emails/utils';
 import { groupIds } from '@blockframes/utils/emails/ids';
-import { User, OrganizationDocument, PublicUser, InvitationDocument, PermissionsDocument, App, ErrorResultResponse } from '@blockframes/model';
+import {
+  getUserEmailData,
+  OrgEmailData,
+  User,
+  OrganizationDocument,
+  PublicUser,
+  InvitationDocument,
+  PermissionsDocument,
+  App,
+  ErrorResultResponse,
+} from '@blockframes/model';
 import { registerToNewsletters, updateMemberTags } from './mailchimp';
 import { getPreferenceTag, MailchimpTag } from '@blockframes/utils/mailchimp/mailchimp-model';
