+++ apps/crm/crm/src/app/crm/components/organization/create-organization/create.component.ts
@@ -3,13 +3,12 @@ import { MatDialogRef } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { OrganizationService } from '@blockframes/organization/service';
 import { Router } from '@angular/router';
-import { createPublicUser, getOrgAppAccess } from '@blockframes/model';
+import { createPublicUser, getOrgAppAccess, OrgEmailData } from '@blockframes/model';
 import { AuthService } from '@blockframes/auth/service';
 import { FormControl, Validators, AbstractControl } from '@angular/forms';
 import { UserService } from '@blockframes/user/service';
 import { FormEntity } from '@blockframes/utils/form';
 import { OrganizationCrmForm } from '@blockframes/admin/crm/forms/organization-crm.form';
-import { OrgEmailData } from '@blockframes/utils/emails/utils';
 import { where } from 'firebase/firestore';
 
 @Component({
