+++ apps/crm/crm/src/app/crm/pages/mails/mails.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/core';
-import { createEmailRequest } from '@blockframes/utils/emails/utils';
+import { createEmailRequest } from '@blockframes/model';
 import { TestEmailForm } from '@blockframes/admin/crm/forms/test-email.form';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { SendgridService } from '@blockframes/utils/emails/sendgrid.service'
