+++ libs/admin/src/lib/crm/forms/test-email.form.ts
@@ -1,6 +1,6 @@
 import { FormControl } from '@angular/forms';
 import { FormEntity } from '@blockframes/utils/form/forms/entity.form';
-import { createEmailRequest, EmailRequest } from '@blockframes/utils/emails/utils';
+import { createEmailRequest, EmailRequest } from '@blockframes/model';
 
 function createTestEmailControls(entity: Partial<EmailRequest>) {
   const request = createEmailRequest(entity);
